Refactor monetary values to use float type
Updated multiple classes to replace decimal with float for monetary properties.
Added PromotionId and StoreId to CreateBookingCommand.
Refactored CreateBookingCommandHandler to generate booking codes and handle new logic for pets and services.
Adjusted validation rules in CreateBookingValidator to reflect data type changes.
Updated database schema in PetDatabaseContextModelSnapshot accordingly.

+++ FurEverCarePlatform.Domain/Entities/Booking.cs
@@ -6,8 +6,8 @@ public class Booking : BaseEntity
     public string? Description { get; set; }
 
     public required Guid UserId { get; set; }
-    public required decimal RawAmount { get; set; }
-    public required decimal TotalAmount { get; set; }
+    public required float RawAmount { get; set; }
+    public required float TotalAmount { get; set; }
     public required string Code { get; set; }
     public Guid? PromotionId { get; set; }
